+++ cypress/e2e/missionZero-wc.cy.js
@@ -95,3 +95,34 @@ it("picks up errors from the editor", () => {
   cy.get("editor-wc").shadow().find(".btn--run").click()
   cy.get("#results").should("contain", '"isErrorFree":false')
 })
+
+it("does not return null duration if no change in focus", () => {
+  cy.get("editor-wc").shadow().find("div[class=cm-content]").invoke('text', 'from sense_hat import SenseHat\nsense = SenseHat()\nsense.send_message("a")')
+  cy.get("editor-wc").shadow().find(".btn--run").click()
+  cy.get('#results').should("not.contain", '"duration":null')
+})
+
+it("does not return null duration if focus changed before code run", () => {
+  cy.get("editor-wc").shadow().find("div[class=cm-content]").invoke('text', 'from sense_hat import SenseHat\nsense = SenseHat()\nsense.send_message("a")')
+  cy.window().blur()
+  cy.window().focus()
+  cy.get("editor-wc").shadow().find(".btn--run").click()
+  cy.get('#results').should("not.contain", '"duration":null')
+})
+
+it("returns duration of null if focus is lost", () => {
+  cy.get("editor-wc").shadow().find("div[class=cm-content]").invoke('text', 'from sense_hat import SenseHat\nsense = SenseHat()\nsense.send_message("a")')
+  cy.get("editor-wc").shadow().find(".btn--run").click()
+  cy.window().blur()
+  cy.window().focus()
+  cy.get('#results').should("contain", '"duration":null')
+})
+
+it("does not return duration of null if code rerun after focus lost", () => {
+  cy.get("editor-wc").shadow().find("div[class=cm-content]").invoke('text', 'from sense_hat import SenseHat\nsense = SenseHat()\nsense.send_message("a")')
+  cy.get("editor-wc").shadow().find(".btn--run").click()
+  cy.window().blur()
+  cy.window().focus()
+  cy.get("editor-wc").shadow().find(".btn--run").click()
+  cy.get('#results').should("not.contain", '"duration":null')
+})