+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Added
+- The web component `runCompleted` event now returns `duration: null` if the host page's tab loses focus during the code run (#192)
+
 ## [0.4.0]
 
 ### Added